+++ source/put/source/ecs/ecs_scene.h
@@ -477,7 +477,7 @@ namespace put
             ecs_controller* controllers = nullptr;
 
             // Scene Data
-            u32                 num_entities = 0;
+            size_t              num_entities = 0;
             u32                 soa_size = 0;
             free_node_list*     free_list_head = nullptr;
             u32                 forward_light_buffer = PEN_INVALID_HANDLE;
